parse_hint: fix function and add unit test
code review indicated that parse_hint() would loop forever when looking
for "stuff=" in "foostuff=bar" (e.g. stuff found not at start of line or
after a comma)

Add test to illustrate the problem and fix it.

+++ copytool/scheduler.c
@@ -88,7 +88,7 @@ schedule_host_mapping_consistent_hash(struct host_mapping *mapping,
 	struct client *client;
 	const char *hostname;
 	const char *value;
-	int value_len;
+	size_t value_len;
 	size_t index;
 	size_t hash;
